# src/main/resources/static/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Moiming - Event Schedule Management System API
  description: 모임 일정 관리 시스템 '모이밍' API 문서
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/events/{publicId}:
    get:
      security: []
      tags:
      - Event
      summary: 이벤트 상세 조회
      description: 이벤트 상세 정보를 조회합니다
      parameters:
      - name: publicId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventDetailResponse"
    put:
      tags:
      - Event
      summary: 이벤트 수정
      description: 이벤트를 수정합니다
      security:
      - bearerAuth: []
      parameters:
      - name: publicId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateEventResponse"
        "401":
          description: 인증 필요
    delete:
      tags:
      - Event
      summary: 이벤트 삭제
      description: 이벤트를 삭제합니다
      security:
      - bearerAuth: []
      parameters:
      - name: publicId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: 인증 필요
...
  /api/events:
    post:
      tags:
      - Event
      summary: 이벤트 생성
      description: 새로운 이벤트를 생성합니다
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateEventResponse"
        "401":
          description: 인증 필요
...
  /api/auth/signup:
    post:
      security: []
      tags:
      - Auth
      summary: 이메일 회원가입
      description: 새로운 사용자를 등록하고 인증 이메일을 발송합니다. 이메일 인증 후 가입이 완료됩니다.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "409":
          description: 이미 존재하는 email
        "204":
          description: 인증 이메일 발송 성공
        "400":
          description: 잘못된 요청 (유효하지 않은 email/password/name)
        "503":
          description: 이메일 서비스 장애
  /api/auth/logout:
    post:
      security:
      - bearerAuth: []
      tags:
      - Auth
      summary: 로그아웃
      description: 현재 JWT 토큰을 블랙리스트에 추가하여 로그아웃합니다
      responses:
        "204":
          description: 로그아웃 성공
  /api/auth/login:
    post:
      security: []
      tags:
      - Auth
      summary: 로그인
      description: email로 로그인하여 JWT 토큰을 발급받습니다
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "401":
          description: 인증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: "로그인 성공, JWT 토큰 반환"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/auth/email-verification/{verificationCode}:
    post:
      security: []
      tags:
      - Auth
      summary: 이메일 인증
      description: 이메일로 받은 인증 코드를 검증하고 회원가입을 완료합니다
      parameters:
      - name: verificationCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: 유효하지 않거나 만료된 인증 코드
        "200":
          description: "인증 성공, 유저 등록 완료"
...
  /api/events/me:
    get:
      tags:
      - Event
      summary: 내가 만든 이벤트 목록 조회 (무한 스크롤)
      description: 로그인된 사용자가 생성한 이벤트 목록을 무한 스크롤 방식(cursor)으로 조회합니다
      security:
      - bearerAuth: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MyEventsInfiniteResponse"
        "401":
          description: 인증 필요
components:
  schemas:
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
          description: 이벤트 제목
          nullable: true
          example: 정기 모임
        description:
          type: string
          description: 이벤트 설명
          nullable: true
          example: 정기 모임입니다
        location:
          type: string
          description: 장소
          nullable: true
          example: 강의실 101
        startsAt:
          type: string
          description: 시작 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T18:00:00Z
        endsAt:
          type: string
          description: 종료 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T20:00:00Z
        capacity:
          type: integer
          description: 정원
          format: int32
          nullable: true
          example: 30
        waitlistEnabled:
          type: boolean
          description: 대기 명단 사용 여부
          nullable: true
          example: true
        registrationStartsAt:
          type: string
          description: 신청 시작 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T17:00:00Z
        registrationEndsAt:
          type: string
          description: 신청 마감 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T17:30:00Z
      description: 일정 수정 요청
    UpdateEventResponse:
      required:
      - registrationEndsAt
      - title
      - capacity
      - waitlistEnabled
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        capacity:
          type: integer
          format: int32
          nullable: true
        waitlistEnabled:
          type: boolean
        registrationStartsAt:
          type: string
          format: date-time
          nullable: true
        registrationEndsAt:
          type: string
          format: date-time
...
    CreateEventRequest:
      required:
      - capacity
      - registrationEndsAt
      - title
      - waitlistEnabled
      type: object
      properties:
        title:
          type: string
          description: 이벤트 제목
          example: 정기 모임
        description:
          type: string
          description: 이벤트 설명
          nullable: true
          example: 정기 모임입니다
        location:
          type: string
          description: 장소
          nullable: true
          example: 강의실 101
        startsAt:
          type: string
          description: 시작 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T18:00:00Z
        endsAt:
          type: string
          description: 종료 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T20:00:00Z
        capacity:
          type: integer
          description: 정원
          format: int32
          example: 30
        waitlistEnabled:
          type: boolean
          description: 대기 명단 사용 여부
          example: true
        registrationStartsAt:
          type: string
          description: 신청 시작 시간 (ISO-8601). null인 경우 요청 시점으로 자동 설정됩니다.
          format: date-time
          nullable: true
          example: 2026-02-02T17:00:00Z
        registrationEndsAt:
          type: string
          description: 신청 마감 시간 (ISO-8601)
          format: date-time
          example: 2026-02-02T17:30:00Z
      description: 일정 생성 요청
    CreateEventResponse:
      required:
      - publicId
      type: object
      properties:
        publicId:
          type: string
...
    LoginResponse:
      required:
      - token
      type: object
      properties:
        token:
          type: string
    SignupRequest:
      required:
      - email
      - name
      - password
      type: object
      properties:
        email:
          type: string
          description: 사용자 이메일
          example: user@example.com
        name:
          type: string
          description: 사용자 이름
          example: 홍길동
        password:
          type: string
          description: 비밀번호
          example: password123
        profileImage:
          type: string
          description: 프로필 이미지 URL
          nullable: true
          example: https://cdn.example.com/profile.png
      description: 회원가입 요청
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
          description: 사용자 이메일
          example: user@example.com
        password:
          type: string
          description: 사용자 비밀번호
          example: mypassword
      description: 로그인 요청
...
    CapabilitiesInfo:
      required:
      - apply
      - cancel
      - shareLink
      - wait
      type: object
      properties:
        shareLink:
          type: boolean
          description: 공유 링크 가능 여부
          example: false
        apply:
          type: boolean
          description: 참여 신청 가능 여부(확정 자리 신청)
          example: false
        wait:
          type: boolean
          description: 대기 신청 가능 여부(정원 찼을 때)
          example: false
        cancel:
          type: boolean
          description: 신청 취소 가능 여부
          example: true
      description: 가능한 동작 블록
    CreatorInfo:
      required:
      - email
      - name
      type: object
      properties:
        name:
          type: string
          description: 이름
          example: 홍길동
        email:
          type: string
          description: 이메일
          example: user@example.com
        profileImage:
          type: string
          description: 프로필 이미지
          nullable: true
          example: https://example.com/profile.png
      description: 생성자 정보 블록
    EventDetailResponse:
      required:
      - capabilities
      - creator
      - event
      - guestsPreview
      - viewer
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventInfo"
        creator:
          $ref: "#/components/schemas/CreatorInfo"
        viewer:
          $ref: "#/components/schemas/ViewerInfo"
        capabilities:
          $ref: "#/components/schemas/CapabilitiesInfo"
        guestsPreview:
          type: array
          description: 참여자 미리보기 목록
          items:
            $ref: "#/components/schemas/GuestPreview"
      description: 일정 상세 조회 응답
    EventInfo:
      required:
      - publicId
      - registrationEndsAt
      - title
      - confirmedCount
      - waitlistCount
      type: object
      properties:
        publicId:
          type: string
          description: 이벤트 publicId
          example: dj8sadj10djnkas
        title:
          type: string
          description: 이벤트 제목
          example: 제2회 기획 세미나
        description:
          type: string
          description: 이벤트 설명
          nullable: true
          example: 일정 상세 설명...
        location:
          type: string
          description: 장소
          nullable: true
          example: 서울대
        startsAt:
          type: string
          description: 시작 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T18:00:00Z
        endsAt:
          type: string
          description: 종료 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T20:00:00Z
        capacity:
          type: integer
          description: 정원
          format: int32
          nullable: true
          example: 10
        confirmedCount:
          type: integer
          description: 참여 확정 신청자 수
          format: int32
          example: 8
        waitlistCount:
          type: integer
          description: 대기자 수
          format: int32
          example: 8
        registrationStartsAt:
          type: string
          description: 신청 시작 시간 (ISO-8601)
          format: date-time
          nullable: true
          example: 2026-02-02T17:00:00Z
        registrationEndsAt:
          type: string
          description: 신청 마감 시간 (ISO-8601)
          format: date-time
          example: 2026-02-02T17:00:00Z
      description: 이벤트 정보 블록
    GuestPreview:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          description: 참여자 ID (비회원이면 null)
          format: int64
          nullable: true
          example: 1
        name:
          type: string
          description: 참여자 이름
          example: 홍길동
        profileImage:
          type: string
          description: 프로필 이미지 URL
          nullable: true
          example: https://example.com/profile.png
      description: 참여자 미리보기 정보
    ViewerInfo:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          description: 이벤트 조회자(Viewer)의 상태
          example: WAITLISTED
          enum:
          - HOST
          - CONFIRMED
          - WAITLISTED
          - BANNED
          - NONE
        name:
          type: string
          description: viewer 이름 (비회원이면 null)
          nullable: true
          example: 홍길동
        waitlistPosition:
          type: integer
          description: "대기 순번 (WAITLISTED일 때만 숫자, 아니면 null)"
          format: int32
          nullable: true
          example: 3
        registrationPublicId:
          type: string
          description: "registrationPublicId (게스트면 값, 아니면 null)"
          nullable: true
          example: "1"
        reservationEmail:
          type: string
          description: "예약 이메일(게스트면 guestEmail, 로그인 유저면 user email 등 정책에 따름)"
          nullable: true
          example: skdfsj@gmaldkl
      description: 조회자(viewer) 정보 블록
...
    MyEventResponse:
      required:
      - publicId
      - registrationEndsAt
      - title
      - confirmedCount
      - waitlistCount
      type: object
      properties:
        publicId:
          type: string
        title:
          type: string
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        registrationStartsAt:
          type: string
          format: date-time
          nullable: true
        registrationEndsAt:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
          nullable: true
        confirmedCount:
          type: integer
          format: int32
        waitlistCount:
          type: integer
          format: int32
    MyEventsInfiniteResponse:
      required:
      - events
      - hasNext
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/MyEventResponse"
        nextCursor:
          type: string
          format: date-time
          nullable: true
        hasNext:
          type: boolean
...
    ErrorResponse:
      required:
      - code
      - message
      - title
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        title:
          type: string
          description: 에러 제목
        message:
          type: string
          description: 에러 메시지
      description: 에러 응답
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT 토큰을 입력하세요
      scheme: bearer
      bearerFormat: JWT